Revision of FMTtheme, fix for infinit recursivity...

+++ Source/FMTmask.cpp
@@ -62,7 +62,7 @@ boost::dynamic_bitset<> FMTmask::subset(const FMTtheme& theme) const
     {
     boost::dynamic_bitset<> sub(theme.size(),false);
     int locit = 0;
-    for(int id = theme.start; id < (theme.start + static_cast<int>(theme.size())); ++id)
+    for(size_t id = theme.start; id < (theme.start + theme.size()); ++id)
         {
 		sub[locit] = data[id];
         ++locit;
@@ -72,7 +72,7 @@ boost::dynamic_bitset<> FMTmask::subset(const FMTtheme& theme) const
 void FMTmask::setsubset(const FMTtheme& theme,const boost::dynamic_bitset<>& subset)
     {
     int locit = 0;
-    for(int id = theme.start; id < (theme.start+static_cast<int>(theme.size())); ++id)
+    for(size_t id = theme.start; id < (theme.start+theme.size()); ++id)
         {
         data[id] = subset[locit];
         ++locit;
